requirements.txt
requests>=2.31.0
feedparser>=6.0.8
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
python-dateutil>=2.8.2
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pytest>=8.0.0
responses>=0.25.0